feat(autoware_tensorrt_vad): add file for Backbone class and Head class
