+++ AV-Solutions/vad-trt/app/autoware_tensorrt_vad/include/autoware/tensorrt_vad/networks/net.hpp
@@ -1,19 +1,16 @@
-/*
- * SPDX-FileCopyrightText: Copyright (c) 2023-2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
- * SPDX-License-Identifier: Apache-2.0
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// Copyright 2025 Shin-kyoto.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 #ifndef NET_H_
 #define NET_H_
@@ -80,37 +77,6 @@ public:
   void Enqueue(cudaStream_t stream);
 
   virtual ~Net();
-};
-
-// BackboneクラスとHeadクラスの定義
-class Backbone : public Net {
-public:
-  Backbone(
-    const VadConfig& vad_config,
-    const autoware::tensorrt_common::TrtCommonConfig& trt_common_config,
-    const std::string& plugins_path,
-    std::shared_ptr<VadLogger> logger
-  );
-
-  std::vector<autoware::tensorrt_common::NetworkIO> generate_network_io(const VadConfig& vad_config) override;
-  void set_input_tensor(TensorMap& ext) override;
-};
-
-class Head : public Net {
-private:
-  NetworkType network_type_;
-
-public:
-  Head(
-    const VadConfig& vad_config,
-    const autoware::tensorrt_common::TrtCommonConfig& trt_common_config,
-    NetworkType network_type,
-    const std::string& plugins_path,
-    std::shared_ptr<VadLogger> logger
-  );
-
-  std::vector<autoware::tensorrt_common::NetworkIO> generate_network_io(const VadConfig& vad_config) override;
-  void set_input_tensor(TensorMap& ext) override;
 };
 
 } // namespace autoware::tensorrt_vad
